FEAT : 경험치 반환 Int로 통일

+++ src/main/kotlin/syntax/backend/runways/dto/ResponseMyInfoDTO.kt
@@ -17,5 +17,5 @@ data class ResponseMyInfoDTO(
     val marketing: Boolean,
     val accountPrivate: Boolean,
     val courses: Page<ResponseCourseDTO>,
-    val experience: Float
+    val experience: Int
 )
