+++ src/main/kotlin/syntax/backend/runways/dto/UserProfileWithCoursesDTO.kt
@@ -10,5 +10,5 @@ data class UserProfileWithCoursesDTO(
     val follow: Follow?,
     val courses: Page<ResponseCourseDTO>,
     val isFollow: Boolean,
-    val experience : Float
+    val experience : Int
 )
